--- Assets/Scripts/Player/Stunnable.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: c206fc2ff8a36634580d44a66cc9894d
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
